+++ IntSorterUtils.h
@@ -12,19 +12,19 @@
 namespace bit_mask_sorter {
     class IntSorterUtils {
     public:
-        static int32_t partitionNotStable(int32_t *array, int32_t start, int32_t end, int32_t mask);
+        static size_t partitionNotStable(int32_t *array, size_t start, size_t end, int32_t mask);
 
-        static int32_t partitionReverseNotStable(int32_t *array, int32_t start, int32_t end, int32_t mask);
+        static size_t partitionReverseNotStable(int32_t *array, size_t start, size_t end, int32_t mask);
 
-        static int32_t partitionStable(int32_t *array, int32_t start, int32_t end, int32_t mask, int32_t *aux);
+        static int32_t partitionStable(int32_t *array, size_t start, size_t end, int32_t mask, int32_t *aux);
 
-        static void partitionStableLastBits(int32_t *array, int32_t start, int32_t end, int32_t mask, int32_t twoPowerK,
+        static void partitionStableLastBits(int32_t *array, size_t start, size_t end, int32_t mask, int32_t twoPowerK,
                                             int32_t *aux);
 
-        static void partitionStableGroupBits(int32_t *array, int32_t start, int32_t end, int32_t mask, int32_t shiftRight,
-                                 int32_t twoPowerK, int32_t *aux);
+        static void partitionStableGroupBits(int32_t *array, size_t start, size_t end, int32_t mask, int32_t shiftRight,
+                                             int32_t twoPowerK, int32_t *aux);
 
-        static void arraycopy(int32_t *pSrc, int srcPos, int32_t *pDest, int32_t destPos, int32_t length);
+        static void arraycopy(int32_t *pSrc, size_t srcPos, int32_t *pDest, size_t destPos, size_t length);
     };
 }
